chore!: upgrade to cdktf 0.18.0 (#169)
This PR initiates the upgrade of CDKTF from version `0.17.0` to version
`0.18.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/provider-null` to a version compatible with
`cdktf@0.18.0`
[here](https://github.com/cdktf/cdktf-local-exec/blob/45a35846c8c7e5bdc600d76344305b1a049ef130/.projenrc.ts#L62)
- [x] Update `@cdktf/provider-random` to a version compatible with
`cdktf@0.18.0`
[here](https://github.com/cdktf/cdktf-local-exec/blob/main/.projenrc.ts#L67)
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Nara Kasbergen <nara.kasbergen@gmail.com>