chore!: upgrade to cdktf 0.19.0 (#182)
This PR initiates the upgrade of CDKTF from version `0.18.0` to version
`0.19.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/provider-null` to a version compatible with
`cdktf@0.19.0`
[here](https://github.com/cdktf/cdktf-local-exec/blob/69b1363d67ed3c23391bbc4f1de24b86b640484f/.projenrc.ts#L64).
Look up the version
[here](https://github.com/cdktf/cdktf-provider-null/releases/).
- [x] Update `@cdktf/provider-random` to a version compatible with
`cdktf@0.19.0`
[here](https://github.com/cdktf/cdktf-local-exec/blob/69b1363d67ed3c23391bbc4f1de24b86b640484f/.projenrc.ts#L67).
Look up the version
[here](https://github.com/cdktf/cdktf-provider-random/releases/).
- [x] Run `npx projen`
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Daniel Schmidt <danielmschmidt92@gmail.com>